Enhances PersistentContainer, Stops spotlight indexing in background
